Fix update shopUrl
